make the headers of Minos and mvm consistent
- changes vmtag
- changes __packed

This may fix that the guest VM can't be created.

Related to #40, #43
