+++ include/minos/compiler.h
@@ -8,7 +8,7 @@
 #define __unused		__attribute__((__unused__))
 #define __align(x)		__attribute__((__aligned__(x)))
 #define __align_cache_line	__align(__cache_line_size__)
-#define __packed__		__attribute__((__packed__))
+#define __packed		__attribute__((__packed__))
 #define __noreturn		__attribute__((noreturn))
 #define likely(x)		__builtin_expect(!!(x), 1)
 #define unlikely(x)		__builtin_expect(!!(x), 0)
