+++ tools/mvm/include/common/hypervisor.h
@@ -8,6 +8,29 @@
 #include <sys/types.h>
 #endif
 
+#define RAMDISK_MAGIC "MINOSRAMDISK...."
+#define RAMDISK_MAGIC_SIZE 16
+
+#define RAMDISK_FNAME_SIZE 32
+
+struct ramdisk_inode {
+	char fname[RAMDISK_FNAME_SIZE];
+	uint64_t f_offset;	// data offset from ramdisk_start.
+	uint64_t f_size;	// data size of this file
+} __packed;
+
+struct ramdisk_sb {
+	uint32_t file_cnt;
+	uint32_t block_size;	// reserved
+	uint64_t inode_offset;
+	uint64_t data_base;	// reserved
+};
+
+struct ramdisk_file {
+	struct ramdisk_inode *inode;
+	unsigned long pos;	// reserved
+};
+
 #define VM_NAME_SIZE	32
 #define VM_TYPE_SIZE	16
 
@@ -39,6 +62,9 @@ struct vmtag {
 	uint64_t setup_data;
 	uint64_t flags;
 	uint32_t vcpu_affinity[8];
+	uint64_t load_address;
+	char image_file[RAMDISK_FNAME_SIZE];
+	char dtb_file[RAMDISK_FNAME_SIZE];
 };
 
 #define IOCTL_CREATE_VM			0xf000
