+++ tools/mvm/include/minos/virtio.h
@@ -62,26 +62,26 @@
 struct vring_used_elem {
 	uint32_t id;
 	uint32_t len;
-} __attribute__((__packed__));
+} __packed;
 
 struct vring_used {
 	uint16_t flags;
 	uint16_t idx;
 	struct vring_used_elem ring[];
-} __attribute__((__packed__));
+} __packed;
 
 struct vring_avail {
 	uint16_t flags;
 	uint16_t idx;
 	uint16_t ring[];
-} __attribute__((__packed__));
+} __packed;
 
 struct vring_desc {
 	uint64_t addr;
 	uint32_t len;
 	uint16_t flags;
 	uint16_t next;
-} __attribute__((__packed__));
+} __packed;
 
 struct virtio_device;
